pyproject.toml
[project]
name = "flowparser"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladimir Karpenko",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "g4f[api] (>=0.5.5.7,<0.6.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "orjson (>=3.10.18,<4.0.0)",
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".idea",
    "migrations",
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
line_length = 120
skip = ["migrations", ".venv", "tests"]
profile = "black"
include_trailing_comma = true
use_parentheses = true
skip_gitignore = true

